イメージの表示部をコンポーネントに切り出す
